rename math helper file
